+++ frontend/src/app/shared/dropdown/dropdown-menu.directive.ts
@@ -4,7 +4,7 @@ import {  takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
 import * as _ from 'lodash';
 import { DropdownService } from './dropdown.service';
-import { modulo } from '../utils/utils';
+import { modulo } from '../../utils/utils';
 
 @Directive({
     selector: '[iaDropdownMenu]'
