+++ frontend/src/app/shared/dropdown/dropdown.component.scss
